requirements.txt
langchain
langchain-community
fpdf2
streamlit
transformers
pdfplumber
faiss-cpu
python-dotenv